Remove the monitor mutex.
Since we only do synchronous scans in the test suite, we remove the
ability to force the monitor to scan and thus remove the need for the
mutex in the monitor itself.
